Reorganise test suite around function
