+++ tests/test_game.py
@@ -97,210 +97,6 @@ def test_from_dict():
     assert pl2 == "b"
 
 
-def test_game_get_outcome():
-    game = gbt.Game.new_table([2, 2])
-    game.make_outcome({"1": "1", "2": "1"}, {"1": 0, "2": 0}, "top left")
-    assert game.get_outcome({"1": "1", "2": "1"}) == next(iter(game.outcomes))
-
-
-def test_game_get_outcome_by_relabeled_strategies():
-    game = gbt.Game.new_table([2, 2])
-    pl1, pl2 = game.players
-    game.relabel_strategies(pl1, {next(iter(game.get_strategies(pl1))): "defect"})
-    game.relabel_strategies(pl2, {next(iter(game.get_strategies(pl2))): "cooperate"})
-    game.make_outcome({pl1: "defect", pl2: "cooperate"}, {"1": 0, "2": 0}, "corner")
-    assert game.get_outcome({pl1: "defect", pl2: "cooperate"}) == \
-        next(iter(game.outcomes))
-
-
-def test_game_get_outcome_incomplete_contingency_raises():
-    game = gbt.Game.new_table([2, 2])
-    with pytest.raises(ValueError):
-        _ = game.get_outcome({"1": "1"})
-
-
-def test_game_get_outcome_unknown_player_raises():
-    game = gbt.Game.new_table([2, 2])
-    with pytest.raises(KeyError):
-        _ = game.get_outcome({"1": "1", "2": "1", "3": "1"})
-
-
-def test_game_get_outcome_non_mapping_raises():
-    game = gbt.Game.new_table([2, 2])
-    with pytest.raises(TypeError):
-        _ = game.get_outcome(42)
-
-
-def test_game_get_outcome_non_str_value_raises():
-    game = gbt.Game.new_table([2, 2])
-    with pytest.raises(TypeError):
-        _ = game.get_outcome({"1": 1.23, "2": "1"})
-
-
-def test_game_get_outcome_unknown_strategy_label_raises():
-    game = gbt.Game.new_table([2, 2])
-    with pytest.raises(KeyError):
-        _ = game.get_outcome({"1": "1", "2": "99"})
-
-
-def test_game_get_outcome_unmatched_label_after_relabel_raises():
-    game = gbt.Game.new_table([2, 2])
-    pl1, pl2 = game.players
-    game.relabel_strategies(pl1, {next(iter(game.get_strategies(pl1))): "defect"})
-    game.relabel_strategies(pl2, {next(iter(game.get_strategies(pl2))): "cooperate"})
-    with pytest.raises(KeyError):
-        _ = game.get_outcome({pl1: "defect", pl2: "defect"})
-
-
-def test_game_get_outcome_tree_raises():
-    game = gbt.Game.new_tree(["Alice"])
-    game.append_move(gbt.H.path(), "Alice", ["a", "b"])
-    with pytest.raises(gbt.UndefinedOperationError):
-        _ = game.get_outcome({"Alice": "a"})
-
-
-def test_game_get_payoffs():
-    game = gbt.Game.new_table([2, 2])
-    game.make_outcome({"1": "1", "2": "1"}, {"1": 3, "2": -3}, "top left")
-    payoffs = game.get_payoffs({"1": "1", "2": "1"})
-    assert payoffs["1"] == 3
-    assert payoffs["2"] == -3
-
-
-def test_game_get_payoffs_tree():
-    game = gbt.Game.new_tree(["Alice"])
-    game.append_move(gbt.H.path(), "Alice", ["a", "b"])
-    selector = gbt.H.path()
-    strategy = next(
-        s for s in game.get_strategies("Alice")
-        if game.get_behavior("Alice", s).get(selector) == "a"
-    )
-    game.make_outcome(gbt.H.path("a"), {"Alice": 1}, "a-outcome")
-    payoffs = game.get_payoffs({"Alice": strategy})
-    assert payoffs["Alice"] == 1
-
-
-def test_mixed_strategy_profile_game_structure_changed_no_tree():
-    game = gbt.Game.from_arrays([[2, 2], [0, 0]], [[0, 0], [1, 1]])
-    profiles = [game.mixed_strategy_profile(rational=b) for b in [False, True]]
-    player = next(iter(game.players))
-    distribution = {s: 0 for s in game.get_strategies(player)}
-    next(iter(game.outcomes))[player] = 3
-    for profile in profiles:
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.copy()
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.liap_value()
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.max_regret()
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.normalize()
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.payoffs
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.player_regrets
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.strategy_regrets
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.strategy_values
-        with pytest.raises(gbt.GameStructureChangedError):
-            # triggers error via __getitem__
-            next(profile.__iter__())
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.__setitem__(player, distribution)
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.set_mixed_strategy(player, distribution)
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.__getitem__(player)
-
-
-def test_mixed_strategy_profile_game_structure_changed_tree():
-    game = games.read_from_file("basic_extensive_game.efg")
-    profiles = [game.mixed_strategy_profile(rational=b) for b in [False, True]]
-    player = next(iter(game.players))
-    game.set_move_actions(gbt.H.path(), ["D1"], drop=True)
-    distribution = {s: 0 for s in game.get_strategies(player)}
-    for profile in profiles:
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.as_behavior()
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.copy()
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.liap_value()
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.max_regret()
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.normalize()
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.payoffs
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.player_regrets
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.strategy_regrets
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.strategy_values
-        with pytest.raises(gbt.GameStructureChangedError):
-            # triggers error via __getitem__
-            next(profile.__iter__())
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.__setitem__(player, distribution)
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.set_mixed_strategy(player, distribution)
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.__getitem__(player)
-
-
-def test_mixed_behavior_profile_game_structure_changed():
-    game = games.read_from_file("basic_extensive_game.efg")
-    profiles = [game.mixed_behavior_profile(rational=b) for b in [False, True]]
-    game.set_move_actions(gbt.H.path(), ["D1"], drop=True)
-    for profile in profiles:
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.action_regrets
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.action_values
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.as_strategy()
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.beliefs
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.copy()
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.infoset_probs
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.infoset_regrets
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.infoset_values
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.agent_liap_value()
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.liap_value()
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.agent_max_regret()
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.max_regret()
-        with pytest.raises(gbt.GameStructureChangedError):
-            # triggers error via __getitem__
-            next(profile.__iter__())
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.node_values
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.normalize()
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.payoffs
-        with pytest.raises(gbt.GameStructureChangedError):
-            _ = profile.realiz_probs
-        with pytest.raises(gbt.GameStructureChangedError):
-            # triggers error via __getitem__
-            next(profile.__iter__())
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.__setitem__(game.root, {})
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.set_mixed_action(game.root, {})
-        with pytest.raises(gbt.GameStructureChangedError):
-            profile.__getitem__(game.root)
-
-
 COLLECTION_GETTERS = [
     pytest.param(lambda g: g.players, id="GamePlayers"),
     pytest.param(lambda g: g.outcomes, id="GameOutcomes"),
