+++ tests/test_strategic_mutations.py
@@ -5,51 +5,6 @@ import pygambit as gbt
 from . import games
 
 
-def test_strategic_game_get_infosets():
-    game = gbt.Game.new_table([2, 2])
-    player, _ = game.players
-    with pytest.raises(gbt.UndefinedOperationError):
-        _ = game.get_infosets(player)
-
-
-def test_strategic_game_root():
-    game = gbt.Game.new_table([2, 2])
-    with pytest.raises(gbt.UndefinedOperationError):
-        _ = game.root
-
-
-def test_strategic_game_nodes():
-    game = gbt.Game.new_table([2, 2])
-    with pytest.raises(gbt.UndefinedOperationError):
-        _ = game.nodes
-
-
-def test_game_behav_profile_error():
-    game = gbt.Game.new_table([2, 2])
-    with pytest.raises(gbt.UndefinedOperationError):
-        _ = game.mixed_behavior_profile()
-
-
-def test_game_is_const_sum():
-    game = games.read_from_file("const_sum_game.nfg")
-    assert game.is_const_sum
-
-
-def test_game_is_not_const_sum():
-    game = games.read_from_file("non_const_sum_game.nfg")
-    assert not game.is_const_sum
-
-
-def test_game_get_min_payoff():
-    game = games.read_from_file("mixed_strategy.nfg")
-    assert game.min_payoff == 0
-
-
-def test_game_get_max_payoff():
-    game = games.read_from_file("mixed_strategy.nfg")
-    assert game.max_payoff == 3
-
-
 def test_relabel_strategies_swap():
     """Swap is well-defined; strategies keep their positions."""
     game = gbt.Game.new_table([2, 2])
